consumer-kit/dataapp-api/modelApi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
mlflow==2.3.1
pydantic==1.9.1
protobuf==3.19.0
boto3==1.21.15
python-multipart
kubernetes
python-dotenv
